fix(orders): change productId type from Long to Integer

+++ src/main/java/com/beutyglam/dto/order/OrderRequestDTO.java
@@ -11,8 +11,8 @@ public class OrderRequestDTO {
     @Data
     public static class OrderItemDTO{
         
-    private Long productId;
+    private Integer productId;
     private Integer quantity;
-    
+
     }
 }
